Additions to Upload Page
Added more features to upload form:
name: a text input for the name of the game.
description: a text input for the description of the game.
exepath: input text, it must be a string that is a path to the executable inside the zip folder.
gbin: a file input, this is the zip file.
thumbnail: a file input, for the thumbnail image.